Refactor models for Policies::StudentLoans
