+++ app/models/policies/student_loans/school_check_email_data_export_presenter.rb
@@ -1,0 +1,17 @@
+module Policies
+  module StudentLoans
+    class SchoolCheckEmailDataExportPresenter
+      include StudentLoans::PresenterMethods
+
+      attr_reader :claim
+
+      def initialize(claim)
+        @claim = claim
+      end
+
+      def subject
+        subject_list(claim.eligibility.subjects_taught).downcase
+      end
+    end
+  end
+end
